parallel NN training + more GPU / worker
