+++ src/models/kerasTF/binary_models.py
@@ -20,7 +20,7 @@ import tensorflow as tf
 from ray.air import session
 # from ray.air.integrations.keras import Callback
 from ray.air.config import ScalingConfig
-from models.kerasTF.models_parallel import train_func_CPU, train_func_GPU, build_model
+from models.kerasTF.models import train_func_CPU, train_func_GPU, build_model
 from ray.air.integrations.keras import ReportCheckpointCallback
 from ray.train.tensorflow import TensorflowTrainer, TensorflowCheckpoint
