+++ src/models/kerasTF/models.py
@@ -109,9 +109,9 @@ class KerasTFModels(ModelsUtils, ABC):
         # Computing variables
         if self._nb_GPU > 0:
             self._use_gpu = True
-            self._n_workers = self._nb_GPU #6
-            self._nb_CPU_per_worker = int(self._nb_CPU_training / self._n_workers) # 4
-            self._nb_GPU_per_worker = 1
+            self._n_workers = self._nb_GPU / 2 # 3
+            self._nb_CPU_per_worker = int(self._nb_CPU_training / self._n_workers) # 8
+            self._nb_GPU_per_worker = 2
         else:
             self._use_gpu = False
             self._n_workers = int(self._nb_CPU_training * 0.2)
@@ -156,12 +156,6 @@ class KerasTFModels(ModelsUtils, ABC):
             ds = ds.materialize()
             datasets[name] = ds
 
-        if self._nb_GPU > 0:
-            self._fit_GPU(datasets)
-        else:
-            self._fit_CPU(datasets)
-
-    def _fit_CPU(self, datasets):
         # Training parameters
         train_params = {
             'batch_size': self.batch_size,
@@ -172,16 +166,22 @@ class KerasTFModels(ModelsUtils, ABC):
             'weights': self._weights
         }
 
+        if self._nb_GPU > 0:
+            train_func = train_func_GPU
+        else:
+            train_func = train_func_CPU
+
         # Define trainer / tuner
         self._trainer = TensorflowTrainer(
-            train_loop_per_worker=train_func_CPU,
+            train_loop_per_worker=train_func,
             train_loop_config=train_params,
             scaling_config=ScalingConfig(
                 trainer_resources={'CPU': self._nb_CPU_data},
                 num_workers=self._n_workers,
                 use_gpu=self._use_gpu,
                 resources_per_worker={
-                    'CPU': self._nb_CPU_per_worker
+                    'CPU': self._nb_CPU_per_worker,
+                    'GPU' : self._nb_GPU_per_worker
                 }
             ),
             run_config=RunConfig(
@@ -193,21 +193,6 @@ class KerasTFModels(ModelsUtils, ABC):
 
         training_result = self._trainer.fit()
         self._model_ckpt = training_result.best_checkpoints[0][0]
-
-    def _fit_GPU(self, datasets):
-        # Training parameters
-        train_params = {
-            'batch_size': self.batch_size,
-            'epochs': self._training_epochs,
-            'size': self._nb_kmers,
-            'nb_cls': self._nb_classes,
-            'taxa': self.taxa,
-            'workdir':self._workdir,
-            'model': self.classifier,
-            'weights': self._weights
-        }
-
-        self._model_ckpt = train_func_GPU(datasets, train_params)
 
     # Models predicting
     #########################################################################################################
@@ -241,40 +226,22 @@ class KerasTFModels(ModelsUtils, ABC):
 
             ds = ds.materialize()
 
-            if self._nb_GPU > 0:
-                probabilities = self._predict_proba_GPU(ds)
-            else:
-                probabilities = self._predict_proba_CPU(ds)
+            self._predictor = BatchPredictor.from_checkpoint(
+                self._model_ckpt,
+                TensorflowPredictor,
+                model_definition = lambda: build_model(self.classifier, self._nb_classes, self._nb_kmers)
+            )
+            predictions = self._predictor.predict(
+                data = ds,
+                feature_columns = [TENSOR_COLUMN_NAME],
+                batch_size = self.batch_size,
+            )
+
+            probabilities = _unwrap_ndarray_object_type_if_needed(predictions.to_pandas()['predictions'])
             
             return probabilities
         else:
             raise ValueError('No data to predict')
-
-    def _predict_proba_CPU(self, ds):
-        print('_predict_proba_CPU')
-        self._predictor = BatchPredictor.from_checkpoint(
-            self._model_ckpt,
-            TensorflowPredictor,
-            model_definition = lambda: build_model(self.classifier, self._nb_classes, self._nb_kmers)
-        )
-        predictions = self._predictor.predict(
-            data = ds,
-            feature_columns = [TENSOR_COLUMN_NAME],
-            batch_size = self.batch_size,
-        )
-
-        probabilities = _unwrap_ndarray_object_type_if_needed(predictions.to_pandas()['predictions'])
-
-        return probabilities
-    
-    def _predict_proba_GPU(self, ds):
-        print('_predict_proba_GPU')
-        model = load_model(self._model_ckpt)
-        probabilities = []
-        for batch in ds.iter_tf_batches(batch_size = self.batch_size):
-            probabilities.extend(model.predict(batch[TENSOR_COLUMN_NAME]))
-
-        return probabilities
 
     @abstractmethod
     def _get_abs_pred(self):
@@ -349,62 +316,62 @@ def train_func_CPU(config):
     gc.collect()
     tf.keras.backend.clear_session()
 
-def train_func_GPU(datasets, config):
+
+def train_func_GPU(config):
     # Parameters
     batch_size = config.get('batch_size', 128)
     epochs = config.get('epochs', 10)
     size = config.get('size')
     nb_cls = config.get('nb_cls')
-    taxa = config.get('taxa')
-    workdir = config.get('workdir')
     model = config.get('model')
     weights = config.get('weights')
 
-    checkpoint = os.path.join(workdir, model)
+    # Model construction
+    strategy = tf.distribute.MirroredStrategy()
+    with strategy.scope():
+        model = build_model(model, nb_cls, size)
 
     # Data
-    train_ds = datasets['train']
-    val_ds = datasets['validation']
+    train_data = session.get_dataset_shard('train')
+    val_data = session.get_dataset_shard('validation')
 
-    # Convert datasets to tensorflow ds & generator
-    train_ds = train_ds.iterator().to_tf(
-        feature_columns = TENSOR_COLUMN_NAME,
-        label_columns = LABELS_COLUMN_NAME,
-        batch_size = batch_size
-    )
-    val_ds = val_ds.to_tf(
-        feature_columns = TENSOR_COLUMN_NAME,
-        label_columns = LABELS_COLUMN_NAME,
-        batch_size = batch_size
-    )
-
-    # Model construction
-    model = build_model(model, nb_cls, size)
-
-    # Callbacks
-    model_file = os.path.join(checkpoint, taxa, '{epoch:03d}.hdf5')
-    model_csv = os.path.join(checkpoint, taxa, 'training_log.csv')
-    modelckpt = ModelCheckpoint(filepath=model_file, monitor='val_loss', save_best_only=True, mode='auto')
-    early = EarlyStopping(monitor='val_loss', mode='auto', patience=10)
-    csv = CSVLogger(model_csv)
-
-    # Training
-    hist = model.fit(
-        train_ds,
-        epochs = epochs,
-        validation_data = val_ds,
-        callbacks = [modelckpt, early, csv],
-        class_weight = weights,
-        verbose = 1
-    )
-
-    # Checkpointing
-    best_model = np.argmin(hist.history['val_loss']) + 1
-    best_model = f'{best_model:03d}.hdf5'
-    best_model = os.path.join(checkpoint, taxa, best_model)
-    
-    return best_model
-
+    for _ in range(epochs):
+        batch_train = train_data.to_tf(
+            feature_columns = TENSOR_COLUMN_NAME,
+            label_columns = LABELS_COLUMN_NAME,
+            batch_size = batch_size,
+            local_shuffle_buffer_size = batch_size,
+            local_shuffle_seed = int(np.random.randint(1,10000, size = 1))
+        )
+        batch_val = val_data.to_tf(
+            feature_columns = TENSOR_COLUMN_NAME,
+            label_columns = LABELS_COLUMN_NAME,
+            batch_size = batch_size,
+            local_shuffle_buffer_size = batch_size,
+            local_shuffle_seed = int(np.random.randint(1,10000, size = 1))
+        )
+        # Training
+        history = model.fit(
+            x = batch_train,
+            validation_data = batch_val,
+            callbacks = [ReportCheckpointCallback()],
+            class_weight = weights,
+            verbose = 0
+        )
+        # Checkpointing
+        session.report({
+            'accuracy': history.history['accuracy'][0],
+            'loss': history.history['loss'][0],
+            'val_accuracy': history.history['val_accuracy'][0],
+            'val_loss': history.history['val_loss'][0],
+        },
+            checkpoint=TensorflowCheckpoint.from_model(model)
+        )
+        gc.collect()
+        tf.keras.backend.clear_session()
+    del model
+    gc.collect()
+    tf.keras.backend.clear_session()
 
 def build_model(classifier, nb_cls, nb_kmers):
     if classifier == 'attention':
